+++ src/interfaces/joke.interface.ts
@@ -1,6 +1,7 @@
 export interface Joke {
-  id: number;
+  id?: number;
   type: string;
   setup: string;
   punchline: string;
+  rating?: number;
 }