+++ src/page/JokesPage.vue
@@ -1,5 +1,6 @@
 <script setup lang="ts">
 import AppHeader from '@/components/organisms/AppHeader.vue';
+import AppJokeList from '@/components/organisms/AppJokeList.vue';
 import AppToolBar from '@/components/organisms/AppToolBar.vue';
 
 </script>
@@ -7,8 +8,15 @@ import AppToolBar from '@/components/organisms/AppToolBar.vue';
 <template>
     <header>
         <AppHeader />
-        <AppToolBar />
     </header>
+    <main class="app-main">
+        <AppToolBar />
+        <AppJokeList />
+    </main>
 </template>
 
-<style scoped></style>
+<style scoped>
+.app-main {
+    padding: 1rem;
+}
+</style>